modify link.

+++ links.py
@@ -27,6 +27,9 @@ def main():
                 continue
             linksrc = os.path.join(root, file)
             linkdst = os.path.join(HOME_DIR, file)
+            if os.path.exists(linkdst):
+                logging.warn('already exists: ' + linkdst)
+                continue
             os.symlink(linksrc, linkdst)
             logging.info('make symlink src={0}, dst={1}'.format(linksrc, linkdst))
 
@@ -35,8 +38,8 @@ def main():
 def exit():
     linksrc = os.path.join(HOME_DIR, '.vimrc')
     linkdst = os.path.join(HOME_DIR, '.config/nvim/init.vim')
-    if not os.path.exists(linkdst):
-        logging.warn('not found file: '+ linkdst)
+    if os.path.exists(linkdst):
+        logging.warn('already exists: '+ linkdst)
         return
 
     # make symlink for neovim.
